requirements.txt
streamlit>=1.26
pytest>=7.4
pillow>=10.0
pytest-mock>=3.10
python-dotenv>=1.0
google-genai>=0.4.0
opencv-python>=4.8
moviepy>=1.3.1
pydub>=0.26
